% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidedeck_classes.R
\name{caption}
\alias{caption}
\alias{caption<-}
\title{Set and retrieve caption for an object}
\usage{
caption(x)

caption(x) <- value
}
\arguments{
\item{x}{any R object}

\item{value}{character, the caption.}
}
\value{
The object x with caption attribute
}
\description{
Get or set a caption attribute on any object. Is used by \link{IQRoutputPPTX} to
select the proper slide layout.
}
\author{
Daniel Kaschek, IntiQuan
}
