% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidedeck_classes.R
\name{clean_IQRoutputSection}
\alias{clean_IQRoutputSection}
\title{Clean the Slide Output Folder}
\usage{
clean_IQRoutputSection(section = NULL, outputFolder = NULL, verbose = FALSE)
}
\arguments{
\item{section}{character or NULL (default), the section to be removed. All sections if NULL. See also section argument of \link{IQRoutputPPTX}.}

\item{outputFolder}{character or NULL (default). The slide folder where the rds files are saved.
By default, slides are saved in the folder \code{getOption("IQSlide.outputfolder")}.}

\item{verbose}{Logical (FALSE by default). Print information during cleaning if TRUE.}
}
\description{
Select a single section to be cleaned or clean all sections from the slide output folder.
}
