% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_with_classes.R
\name{IQ_table}
\alias{IQ_table}
\alias{IQ_table.IQRtable}
\alias{IQ_table.matrix}
\alias{IQ_table.data.frame}
\alias{IQ_table.flextable}
\title{Table output for IQRoutputPPTX}
\usage{
IQ_table(x, ...)

\method{IQ_table}{IQRtable}(x, ...)

\method{IQ_table}{matrix}(x, ...)

\method{IQ_table}{data.frame}(x, ...)

\method{IQ_table}{flextable}(x, ...)
}
\arguments{
\item{x}{R object (matrix, IQRtable, data.frame, flextable)}

\item{...}{currently not used}
}
\value{
ggplot object (any table will be printed as image)
}
\description{
Prepares objects which can be interpreted as tables (matrix, data.frame, flextable) for
output via IQRoutputPPTX.
}
