% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidedeck_classes.R
\name{IQRoutputPPTX}
\alias{IQRoutputPPTX}
\title{Push contents into a PPTX output}
\usage{
IQRoutputPPTX(..., section = NULL, title = NULL, layout = NULL,
  filename, outputFolder = NULL, verbose = TRUE)
}
\arguments{
\item{...}{The contents to be presented on a slide. Currently supported are one or two objects
of type plot or table.}

\item{section}{character, optional section name. Different section names will generate different
subfolders of the slide output folder.}

\item{title}{character, optional slide title.}

\item{layout}{character or NULL (default), currently supported are "Title and Content" (one object) and
"Two Content" (two objects). If \code{NULL}, the layout is chosen automatically according to the number of
arguments passed via \code{...}.}

\item{filename}{character, filename of the rds file.}

\item{outputFolder}{character or NULL (default). The slide folder where the rds files are saved.
By default, slides are saved in the folder \code{.OUTPUTFOLDER_SLIDES} as defined in IQR setup file,
see \link[IQRtools:setup_IQRtools]{IQRtools::setup_IQRtools}'().}

\item{verbose}{logical, if \code{TRUE} (default) show verbose information.}
}
\description{
The function pushes a plot or table with additional information like
title and section into an rds file within a slide folder.
From that folder, a slide deck can be generated using the
\url{IQSlidedeck} function.
}
\seealso{
Other Slidedeck functions: \code{\link{IQSlidedeck}}
}
\author{
Daniel Kaschek, IntiQuan
}
\concept{Slidedeck functions}
