% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidedeck_classes.R
\name{IQSlidedeck}
\alias{IQSlidedeck}
\title{Build a slide deck from a slide folder}
\usage{
IQSlidedeck(title = NULL, subtitle = NULL, affiliation = NULL,
  date = NULL, filename = "slides.pptx", section = NULL,
  rdspath = NULL, template = NULL)
}
\arguments{
\item{title}{character or NULL (default), title of the slide deck.}

\item{subtitle}{character or NULL (default), subtitle of the slide deck.}

\item{affiliation}{character or NULL (default), author name and company.}

\item{date}{character or NULL (default), date.}

\item{filename}{character, output file name of the slide deck. If a single file name, the output is
generated in rdspath. If the filename contains path information, the file will be generated in the
exact filename location.}

\item{section}{character or \code{NULL} (default), allows to create slides for a specific section only.}

\item{rdspath}{character or \code{NULL} (default), path to the slide files. Searches for slides
in \code{.OUTPUTFOLDER_SLIDES} if \code{NULL}, see \url{IQRtools::setup_IQRtools}.}

\item{template}{character or \code{NULL} (default), path to the template PPTX file. Uses the
internal template if \code{NULL}.}
}
\description{
Reads all rds files and generates a slide deck from the input
}
\details{
If non of the title-relevant arguments are supplied, a slide deck without
title page is produced.
}
\seealso{
Other Slidedeck functions: \code{\link{IQRoutputPPTX}}
}
\author{
Daniel Kaschek, IntiQuan
}
\concept{Slidedeck functions}
