% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidedeck_classes.R
\name{select_plot}
\alias{select_plot}
\title{Subsetting IQRoutputFigures}
\usage{
select_plot(figure, n)
}
\arguments{
\item{figure}{IQRoutputFigure}

\item{n}{integer or range indicating the plots to keep in the figure}
}
\value{
IQRoutputFigure object
}
\description{
Subsetting IQRoutputFigures
}
