% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidedeck_classes.R
\name{IQSlidedeck}
\alias{IQSlidedeck}
\title{Build a slide deck from a slide folder}
\usage{
IQSlidedeck(title = NULL, subtitle = NULL, affiliation = NULL,
  date = NULL, filename = "slides.pptx", section = NULL,
  rdspath = NULL, template = NULL)
}
\arguments{
\item{title}{character or NULL (default), title of the slide deck.}

\item{subtitle}{character or NULL (default), subtitle of the slide deck.}

\item{affiliation}{character or NULL (default), author name and company.}

\item{date}{character or NULL (default), date.}

\item{filename}{character, output file name of the slide deck. If a single file name, the output is
generated in rdspath. If the filename contains path information, the file will be generated in the
exact filename location.}

\item{section}{character or \code{NULL} (default), allows to create slides for a specific section only.}

\item{rdspath}{character or \code{NULL} (default), path to the slide files. Searches for slides
in \code{.OUTPUTFOLDER_SLIDES} if \code{NULL}, see \url{IQRtools::setup_IQRtools}.}

\item{template}{character or \code{NULL} (default), path to the template PPTX file. Uses the
internal template if \code{NULL}.}
}
\description{
Reads all rds files and generates a slide deck from the input
}
\details{
If non of the title-relevant arguments are supplied, a slide deck without
title page is produced.
}
\examples{
\dontrun{


library(IQRtools)
library(ggplot2)
library(flextable)
setwd(tempdir())

# For compliance information in the footer of the slides
IQRinitCompliance("IQSlidedeck.R")

# ------------------------------------------------------
# Slide 1: Create a plot and a bullet point list
# ------------------------------------------------------

p1 <- IQRggplot(cars, aes(x = speed, y = dist)) + geom_point()
caption(p1) <- "Source: Ezekiel, M. (1930) Methods of Correlation Analysis. Wiley."

IQRoutputPPTX(
  c("Cars data set from R datasets",
    "The data vie the speed of cars and distances taken to stop shown as plot"),
  p1,
  section = "Slides with plots",
  title = "Overview of cars distance vs time to stop",
  filename = "carsPlot"
)


# ------------------------------------------------------
# Slide 2: Create a table and a bullet point list
# ------------------------------------------------------

table1 <- flextable(cars[1:16,])

IQRoutputPPTX(
  c("Cars data set from R datasets",
    "The data vie the speed of cars and distances taken to stop"),
  table1,
  section = "Slides with tables",
  title = "Overview of cars distance vs time to stop shown as table",
  filename = "carsTable"
)

# ------------------------------------------------------
# Slide 3: Create a table and a formatted bullet point list
# ------------------------------------------------------

table1 <- flextable(cars[1:16,])

IQRoutputPPTX(
  bullet_list(
    "* Bullet point with **bold** and *italic*" ,
    "* Another bullet point",
    "  * Indented by two spaces",
    "  * Another one",
    "* And back to ***really* outer** level",
    "* Some sort of formula `E_rel_ = m\\\\*c^2^`"
  ),
  table1,
  section = "Slides with tables",
  title = "Overview of cars distance vs time to stop shown as table",
  filename = "carsTable2"
)



# ------------------------------------------------------
# Finally: Create
# ------------------------------------------------------
# Copy-paste the filename shown in R console
# to explorer to open the file

# All slides
IQSlidedeck(
  title = "My first slidedeck with IQSlides",
  subtitle = "Cars and the time they take to stop",
  affiliation = "Daniel Kaschek, IntiQuan",
  date = "2019-11-20"
)


# Just section slides
IQSlidedeck(section = "Slides with plots")

}
}
\seealso{
Other Slidedeck functions: \code{\link{IQRoutputPPTX}}
}
\author{
Daniel Kaschek, IntiQuan
}
\concept{Slidedeck functions}
